Move "D59" into the GEMGeometry/test folder, remove changes to dict2026Geometry
