+++ Configuration/Geometry/python/dict2026Geometry.py
@@ -904,7 +904,7 @@ muonDict = {
         "era" : "phase2_muon, run3_GEM",
     },
 
-    "M6" : {
+    "M5" : {
         1 : [
             'Geometry/MuonCommonData/data/mbCommon/2021/v1/mbCommon.xml',
             'Geometry/MuonCommonData/data/mb1/2015/v2/mb1.xml',
@@ -912,28 +912,28 @@ muonDict = {
             'Geometry/MuonCommonData/data/mb3/2015/v2/mb3.xml',
             'Geometry/MuonCommonData/data/mb4/2015/v2/mb4.xml',
             'Geometry/MuonCommonData/data/mb4Shield/2021/v1/mb4Shield.xml',
-            'Geometry/MuonCommonData/data/muonYoke/2021/v2/muonYoke.xml',
-            'Geometry/MuonCommonData/data/csc/2021/v1/csc.xml',
-            'Geometry/MuonCommonData/data/mfshield/2017/v1/mfshield.xml',
-            'Geometry/MuonCommonData/data/mf/2026/v4/mf.xml',
-            'Geometry/MuonCommonData/data/rpcf/2026/v2/rpcf.xml',
+            'Geometry/MuonCommonData/data/muonYoke/2021/v3/muonYoke.xml',
+            'Geometry/MuonCommonData/data/csc/2021/v2/csc.xml',
+            'Geometry/MuonCommonData/data/mfshield/2017/v2/mfshield.xml',
+            'Geometry/MuonCommonData/data/mf/2026/v3/mf.xml',
+            'Geometry/MuonCommonData/data/rpcf/2026/v3/rpcf.xml',
             'Geometry/MuonCommonData/data/gemf/TDR_BaseLine/gemf.xml',
             'Geometry/MuonCommonData/data/gem11/TDR_BaseLine/gem11.xml',
             'Geometry/MuonCommonData/data/gem21/TDR_Dev/gem21.xml',
-            'Geometry/MuonCommonData/data/ge0/TDR_Dev/v2/ge0.xml',
             'Geometry/MuonCommonData/data/mfshield/2026/v1/mfshield.xml',
+            'Geometry/MuonCommonData/data/me0/TDR_Dev/v2/me0.xml',
         ],
         2 : [
-            'Geometry/MuonCommonData/data/muonNumbering/TDR_DeV/v2/muonNumbering.xml',
+            'Geometry/MuonCommonData/data/muonNumbering/TDR_DeV/v1/muonNumbering.xml',
         ],
         3 : [
-            'Geometry/MuonSimData/data/PhaseII/v2/muonSens.xml',
+            'Geometry/MuonSimData/data/PhaseII/ME0EtaPart/muonSens.xml',
             'Geometry/DTGeometryBuilder/data/dtSpecsFilter.xml',
             'Geometry/CSCGeometryBuilder/data/cscSpecsFilter.xml',
             'Geometry/CSCGeometryBuilder/data/cscSpecs.xml',
             'Geometry/RPCGeometryBuilder/data/2026/v1/RPCSpecs.xml',
-            'Geometry/GEMGeometryBuilder/data/v9/GEMSpecsFilter.xml',
-            'Geometry/GEMGeometryBuilder/data/v9/GEMSpecs.xml',
+            'Geometry/GEMGeometryBuilder/data/v7/GEMSpecsFilter.xml',
+            'Geometry/GEMGeometryBuilder/data/v7/GEMSpecs.xml',
         ],
         4 : [
             'Geometry/MuonSimData/data/PhaseII/muonProdCuts.xml',
@@ -942,6 +942,7 @@ muonDict = {
             'from Geometry.MuonNumbering.muonNumberingInitialization_cfi import *',
             'from RecoMuon.DetLayers.muonDetLayerGeometry_cfi import *',
             'from Geometry.GEMGeometryBuilder.gemGeometry_cfi import *',
+            'from Geometry.GEMGeometryBuilder.me0Geometry_cfi import *',
             'from Geometry.CSCGeometryBuilder.idealForDigiCscGeometry_cff import *',
             'from Geometry.DTGeometryBuilder.idealForDigiDtGeometry_cff import *',
         ],
@@ -1200,7 +1201,7 @@ detectorVersionDict = {
     ("O4","T18","C9","M4","F2","I10") : "D52",
     ("O4","T15","C9","M4","F2","I12") : "D53",
     ("O4","T19","C9","M4","F2","I10") : "D54",
-    ("O4","T19","C9","M6","F2","I10") : "D59",
+    ("O4","T19","C9","M5","F2","I10") : "D55",
  }
 
 deprecatedDets = set([ "D1", "D2", "D3", "D5", "D6" , "D7", "D4", "D8" , "D9", "D12", "D13", "D15", "D10", "D11", "D14", "D16", "D17", "D18", "D19", "D20", "D21", "D22", "D23", "D24", "D25", "D26", "D27", "D28", "D29", "D30", "D31", "D32", "D33", "D34", "D36", "D37", "D38", "D39", "D40", "D42" ])
