+++ Geometry/GEMGeometry/test/GeometryExtended2026GE0Test_cff.py
@@ -3,7 +3,7 @@ import FWCore.ParameterSet.Config as cms
 # This config was generated automatically using generate2026Geometry.py
 # If you notice a mistake, please update the generating script, not just this config
 
-from Geometry.CMSCommonData.cmsExtendedGeometry2026D59XML_cfi import *
+from cmsExtendedGeometry2026GE0TestXML_cfi import *
 from Geometry.TrackerNumberingBuilder.trackerNumberingGeometry_cfi import *
 from SLHCUpgradeSimulations.Geometry.fakeConditions_phase2TkT14_cff import *
 from Geometry.EcalCommonData.ecalSimulationParameters_cff import *
